backend/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
dotenv==0.9.9
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
multidict==6.1.0
packaging==24.2
pillow==11.1.0
propcache==0.3.0
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.0.1
razorpay==1.4.2
requests==2.32.3
setuptools==75.8.0
sqlparse==0.5.3
twilio==9.4.6
urllib3==2.3.0
whitenoise==6.9.0
yarl==1.18.3
